moving bussiness logic from frontEnd, switching to bottom sheet for search

+++ shared/src/commonMain/kotlin/kmp/shared/domain/usecase/trip/DeleteTripUseCase.kt
@@ -1,5 +1,6 @@
 package kmp.shared.domain.usecase.trip
 
+import kmp.shared.base.ErrorResult
 import kmp.shared.base.Result
 import kmp.shared.base.usecase.UseCaseResult
 import kmp.shared.domain.model.Trip
@@ -17,10 +18,16 @@ internal class DeleteTripUseCaseImpl internal constructor(
     private val photoRepository: PhotoRepository
 ): DeleteTripUseCase {
     override suspend fun invoke(params: Trip): Result<Unit> {
-        tripRepository.deleteTripById(params.id)
-        placeRepository.deleteByTripId(params.id)
-        distanceRepository.deleteDistancesByTripId(params.id)
-        photoRepository.deletePhotoByTripId(params.id)
+        val trip = tripRepository.deleteTripById(params.id)
+        if(trip is Result.Error) return trip
+
+        val place = placeRepository.deleteByTripId(params.id)
+        if(place is Result.Error) return place
+
+        val distance = distanceRepository.deleteDistancesByTripId(params.id)
+        if(distance is Result.Error) return distance
+        val photo = photoRepository.deletePhotoByTripId(params.id)
+        if(photo is Result.Error) return photo
 
         return Result.Success(Unit)
     }
